requirements.txt
# Common
numpy>=1.20
PyQt5>=5.15
paho-mqtt>=1.6
cryptography>=3.0

# Canon camera (gphoto2)
gphoto2-cffi>=0.4
Pillow>=9.0
termcolor>=2.0
opencv-python>=4.5

# SPTT camera (CSDU-429)
pyusb>=1.2
astropy>=5.0
